% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispFluorSimilarity.R
\name{dispFluorSimilarity}
\alias{dispFluorSimilarity}
\title{Displays a heatmap of similarity indices for fluorochromes in the panel}
\usage{
dispFluorSimilarity(panel)
}
\arguments{
\item{panel}{list of fluors as characters strings}
}
\description{
Displays a heatmap of similarity indices for fluorochromes in the panel
}
